requirements.txt
numpy==1.26.4
streamlit==1.31.1
opencv-python==4.9.0.80
requests==2.31.0
Pillow==10.1.0
scikit-image==0.22.0